Fix warning
(cherry picked from commit 848fe8811919022d3a21419dc547ab7fd25003ad)

+++ source/app/rendering/graphrenderer.cpp
@@ -888,7 +888,8 @@ void GraphRenderer::processEventQueue()
 
         case QEvent::Type::NativeGesture:
             _interactor->nativeGestureEvent(nativeEvent->gestureType(),
-                nativeEvent->pos() * _devicePixelRatio, nativeEvent->value());
+                nativeEvent->pos() * _devicePixelRatio,
+                static_cast<float>(nativeEvent->value()));
 
             break;
